- inserted new library tests inside DidYouMean - some redesign of DidYouMean that was necessary to follow   a special rule how a library should be used:   - the library provides words that start or end with a test     word which may be possibly also an empty set of words   - all words that the DidYouMean produced with the four     production rules are used to generate a set of     library-completed words   - if this process results in any words from the library,     only library-genrated words are taken   - if the is no library-generated word at all, take the     artifial generated word   - all words that result from these rules are tested against     the index   - the result is ordered using a lightweight comparator that     prefers short words   - a not-so-much-io test against the index is beeing prepared     next - insered the library initialization into the switchboard
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@6284 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/kelondro/index/RowCollection.java
@@ -219,7 +219,7 @@ public class RowCollection implements Iterable<Row.Entry> {
      * @return
      */
     public final long memoryNeededForGrow() {
-        return (long) ((((long) (chunkcount + 1)) * ((long) rowdef.objectsize)) * growfactor100 / 100L);
+        return (((long) (chunkcount + 1)) * ((long) rowdef.objectsize)) * growfactor100 / 100L;
     }
     
     public synchronized void trim(final boolean plusGrowFactor) {
